refactor(tests): split FlowOrchestratorService branch validation tests
Extract branch validation and baseline snapshot tests from
test_flow_orchestrator_service.py into dedicated test_flow_orchestrator_branch_baseline.py.

This split addresses CI LOC limit for test_flow_orchestrator_service.py
(802→674 lines, under 700 limit) while improving test organization.

Extracted tests (8 total):
- Branch validation: test_validate_or_recreate_branch_* (3 tests)
- Creation source recording: test_bootstrap_issue_flow_*_creation_source (3 tests)
- Baseline snapshot: test_bootstrap_issue_flow_*_baseline (2 new tests for force_baseline)

New test_flow_orchestrator_branch_baseline.py (188 lines) covers:
- Branch validation logic (_validate_or_recreate_branch behavior)
- Creation source tracking (when scene_base_ref is recorded)
- Baseline snapshot creation (default idempotent + force override)

Refs: Issue #2762
